refactor: removed unused files
